requirements.txt
python-dotenv~=1.1.0
langchain-openai~=0.3.8
langgraph~=0.3.18
langchain-core~=0.3.47
typing_extensions~=4.12.2
langchain-community~=0.3.20
langchain-text-splitters~=0.3.7
pypdf~=5.4.0